refactor : 댓글/대댓글 조회 시 좋아요 갯수 표시
