+++ src/comment/comment.module.ts
@@ -6,7 +6,7 @@ import { Comment } from './entities/comment.entity';
 import { CommentMapper } from './mapper/comment.mapper';
 import { CommentLike } from './entities/comment-like.entity';
 import { User } from '../user/entities/user.entity';
-import { CommentLikeMapper } from './mapper/commentLike.mapper';
+import { CommentLikeMapper } from './mapper/comment-like.mapper';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Comment, User, CommentLike])],
